+++ test/cheerio_crawler.js
@@ -231,7 +231,7 @@ describe('CheerioCrawler', () => {
                 const opts = crawler._getRequestOptions(request);
                 headers.push(opts.headers);
                 // it needs to return something valid
-                return Object.assign({ body: 'html' }, responseMock);
+                return { dom: {}, response: responseMock };
             };
 
             await crawler.run();
@@ -383,7 +383,7 @@ describe('CheerioCrawler', () => {
                 const opts = crawler._getRequestOptions(request);
                 headers.push(opts.headers);
                 // it needs to return something valid
-                return Object.assign({ body: 'html' }, responseMock);
+                return { dom: {}, response: responseMock };
             };
 
             await crawler.run();
@@ -679,7 +679,7 @@ describe('CheerioCrawler', () => {
                 const opts = crawler._getRequestOptions(request);
                 proxies.push(opts.proxy);
                 // it needs to return something valid
-                return Object.assign({ body: 'html' }, responseMock);
+                return { dom: {}, response: responseMock };
             };
 
             const shuffled = crawler.proxyUrls;
@@ -709,7 +709,7 @@ describe('CheerioCrawler', () => {
                 const opts = crawler._getRequestOptions(request);
                 proxies.push(opts.proxy);
                 // it needs to return something valid
-                return Object.assign({ body: 'html' }, responseMock);
+                return { dom: {}, response: responseMock };
             };
 
             await crawler.run();
@@ -746,7 +746,7 @@ describe('CheerioCrawler', () => {
                 const opts = crawler._getRequestOptions(request);
                 proxies.push(opts.proxy);
                 // it needs to return something valid
-                return Object.assign({ body: 'html' }, responseMock);
+                return { dom: {}, response: responseMock };
             };
 
             await crawler.run();